requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
groq==0.9.0
Pillow==10.3.0
python-dotenv==1.0.1
gunicorn==20.1.0
python-docx==1.1.2